Add indefinite retransmit (retransmit_count: -1)
A publish() with retransmit_count = -1 keeps emitting the same datagram
at retransmit_delay intervals until another publish() to the same topic
supersedes it. Useful for "occupancy/state" style topics where missing
the only-ever-sent packet would leave subscribers wrong indefinitely.

Implementation: cancellable self-rescheduling timeout keyed on
topic_crc32. The std::function reschedules itself; lifecycle is owned
by indefinite_jobs_[topic] so that erasing the map entry + cancel_timeout
breaks the chain. Migrated to set_timeout(uint32_t id, ...) /
cancel_timeout(uint32_t id) since the std::string overload is deprecated
(removed in ESPHome 2026.7).

Config:
- retransmit_count widened to int16_t (1..255 or -1).
- 0 is explicitly rejected.
- -1 requires retransmit_delay >= 1s (cross-field check in
  FINAL_VALIDATE_SCHEMA) so a misconfigured indefinite chain can't flood
  the link.
- Per-action override accepts -1 too; runtime uses the component's
  configured delay.

A new publish() to the same topic always cancels any prior indefinite
chain for that topic, regardless of the new publish's retransmit_count.
Finite retransmit chains aren't tracked and run independently.

+++ components/multicast_pubsub/automation.h
@@ -18,8 +18,9 @@ template<typename... Ts> class PublishAction : public Action<Ts...> {
   TEMPLATABLE_VALUE(std::string, topic)
   TEMPLATABLE_VALUE(std::string, payload)
   // Per-call override of the component-level retransmit_count. If not
-  // set the parent's configured default is used.
-  TEMPLATABLE_VALUE(uint8_t, retransmit_count)
+  // set the parent's configured default is used. -1 means indefinite
+  // (subsequent publish() to the same topic supersedes it).
+  TEMPLATABLE_VALUE(int16_t, retransmit_count)
 
   void play(Ts... x) override {
     auto t = this->topic_.value(x...);
@@ -30,7 +31,7 @@ template<typename... Ts> class PublishAction : public Action<Ts...> {
     // and the set_timeout lambdas only re-send the captured datagram --
     // they don't re-read the count -- so restoring afterwards is safe.
     if (this->retransmit_count_.has_value()) {
-      uint8_t saved = this->parent_->get_retransmit_count();
+      int16_t saved = this->parent_->get_retransmit_count();
       this->parent_->set_retransmit_count(this->retransmit_count_.value(x...));
       this->parent_->publish(t, p);
       this->parent_->set_retransmit_count(saved);
